% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{add_exon_frames}
\alias{add_exon_frames}
\title{Add frames to exons dependent on cds location}
\usage{
add_exon_frames(gtf_df)
}
\arguments{
\item{gtf_df}{A \code{data.frame} or \code{data.table} containing
GTF annotation data, typically from [load_gtf_long()].}
}
\value{
A \code{data.table} with added start_frame and stop_frame
}
\description{
Internal helper
}
\keyword{internal}
