% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarized_experiment_utils.R
\name{add_splice_part}
\alias{add_splice_part}
\title{Add one part to an existing SpliceImpactResult}
\usage{
add_splice_part(obj, data = NULL, res = NULL, hits_final = NULL)
}
\arguments{
\item{obj}{`SpliceImpactResult`}

\item{data}{Optional raw sample-level table.}

\item{res}{Optional differential result table.}

\item{hits_final}{Optional paired/final hits table.}
}
\value{
Updated `SpliceImpactResult`
}
\description{
Add one part to an existing SpliceImpactResult
}
