% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameShiftDetector.R
\name{alheRead}
\alias{alheRead}
\title{Get alignment status of last or hybrid last exons}
\usage{
alheRead(addInf, et, coding_exons, exon_data, exon_length_df, newgtf)
}
\arguments{
\item{addInf}{addInf column passed -- extra info}

\item{et}{exon_type}

\item{coding_exons}{coding exons from initFrameShift}

\item{exon_length_df}{exon_length_df from initFrameShift}

\item{newgtf}{from setup_gtf}

\item{exon_length}{exon_data from initFrameShift}
}
\value{
a character vector of types of alignmetns of each pair
}
\description{
Get alignment status of last or hybrid last exons
}
\keyword{internal}
