% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameShiftDetector.R
\name{alignmentScorer}
\alias{alignmentScorer}
\title{score alignments of proteins}
\usage{
alignmentScorer(type, proBed)
}
\arguments{
\item{type}{event type}

\item{proBed}{proBed paired output}
}
\value{
align scores and lengths
}
\description{
score alignments of proteins
}
\keyword{internal}
