% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarized_experiment_utils.R
\name{as_granges_res}
\alias{as_granges_res}
\title{Convert differential table to GRanges}
\usage{
as_granges_res(res_dt)
}
\description{
Convert differential table to GRanges
}
\keyword{internal}
