% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarized_experiment_utils.R
\name{as_se_raw_events}
\alias{as_se_raw_events}
\title{Convert sample-level table to SummarizedExperiment}
\usage{
as_se_raw_events(data_dt)
}
\description{
Convert sample-level table to SummarizedExperiment
}
\keyword{internal}
