% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarized_experiment_utils.R
\name{as_segments_grl}
\alias{as_segments_grl}
\title{Convert paired hit segments to GRangesList}
\usage{
as_segments_grl(hits_dt)
}
\description{
Convert paired hit segments to GRangesList
}
\keyword{internal}
