% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{ASmatcher}
\alias{ASmatcher}
\title{specific matcher for A5SS/A3SS}
\usage{
ASmatcher(
  i,
  below_thresh = 0.2,
  redExon = redExon,
  minOverlap = minOverlap,
  gtf_transcripts = gtf_transcripts,
  gtf_exons = gtf_exons,
  protein_coding_transcripts = protein_coding_transcripts,
  transcript_starts = transcript_starts
)
}
\description{
specific matcher for A5SS/A3SS
}
