% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedify.R
\name{bedifyBackground}
\alias{bedifyBackground}
\title{Get bed file format from matched transcripts / exons / events
for foreground and background separately}
\usage{
bedifyBackground(matched, outname, cores, gtf)
}
\arguments{
\item{matched}{from matcher function}

\item{cores}{# of cores requested}

\item{gtf}{gtf dataframe from setup_gtf}

\item{output_location}{location where everything is being saved}
}
\value{
a bed file
}
\description{
Get bed file format from matched transcripts / exons / events
for foreground and background separately
}
\keyword{internal}
