% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jaccard.R
\name{calculate_jaccard}
\alias{calculate_jaccard}
\title{calculates jaccard between sequences}
\usage{
calculate_jaccard(start1, stop1, start2, stop2)
}
\value{
jaccard index and alt jacc length
}
\description{
calculates jaccard between sequences
}
\keyword{internal}
