% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{check_extdata_dir}
\alias{check_extdata_dir}
\title{Helper to fetch test dirs for vignettes / tests}
\usage{
check_extdata_dir(dir_name, package = "SpliceImpactR", error = TRUE)
}
\arguments{
\item{dir_name}{Name (or relative path) of the directory inside `inst/extdata/`.}

\item{package}{Package name (defaults to `"SpliceImpactR"`).}

\item{error}{Logical; if TRUE (default), stops with an informative message}
}
\value{
proper directory for data
}
\description{
Helper to fetch test dirs for vignettes / tests
}
\examples{
check_extdata_dir("rawData")

}
\keyword{internal}
