% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarized_experiment_utils.R
\name{coerce_to_dt}
\alias{coerce_to_dt}
\title{Coerce input to data.table for internal pipelines}
\usage{
coerce_to_dt(
  x,
  what = c("raw_events", "di_events", "res_di", "matched", "sample_frame",
    "hits_sequences", "paired_hits")
)
}
\description{
Coerce input to data.table for internal pipelines
}
\keyword{internal}
