% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domains.R
\name{collapse_domains}
\alias{collapse_domains}
\title{Internal helper that collapses a character vector of domain identifiers into
a single, stable | delimited string.}
\usage{
collapse_domains(v)
}
\arguments{
\item{v}{Character vector of domain identifiers.}
}
\value{
Single string joining unique sorted identifiers.
}
\description{
Internal helper that collapses a character vector of domain identifiers into
a single, stable | delimited string.
}
\keyword{internal}
