% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_inclusion_HITindex.R
\name{differential_inclusion_HITindex}
\alias{differential_inclusion_HITindex}
\title{Extract differentially included AFE, ALE from HIT Index data}
\usage{
differential_inclusion_HITindex(
  test_names,
  control_names,
  et,
  outlier_threshold = c("4/n", "1", "Inf")[3],
  minReads = 10,
  min_prop_samples = 0.5,
  chosen_method
)
}
\arguments{
\item{test_names}{a vector of test_names}

\item{control_names}{a vector of control_names}

\item{et}{string of the exon_type}

\item{outlier_threshold}{the thresholding of the cooks distance, no outlier removal is "Inf"}

\item{minReads}{threshold to count an exon as present}

\item{min_prop_samples}{minimum proportion of samples in either phenotype an exon has to be present in to be kept}

\item{chosen_method}{stats model}

\item{cores}{the number of cores requested}
}
\value{
a dataframe with differential inclusion information
}
\description{
Extract differentially included AFE, ALE from HIT Index data
}
