% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_inclusion_rMATS.R
\name{differential_inclusion_rMATS}
\alias{differential_inclusion_rMATS}
\title{Extract differentially included AS events from HIT Index data}
\usage{
differential_inclusion_rMATS(
  control_names,
  test_names,
  et,
  outlier_threshold = c("4/n", "1", "Inf")[3],
  minReads = 10,
  min_prop_samples = 0.5,
  chosen_method
)
}
\arguments{
\item{control_names}{a vector of control_names}

\item{test_names}{a vector of test_names}

\item{et}{string of the exon_type}

\item{outlier_threshold}{the thresholding of the cooks distance, no outlier removal is "Inf"}

\item{minReads}{read num to threshold on}

\item{min_prop_samples}{min prop of samples to require an event to be identified in}

\item{chosen_method}{method for stats analysis}
}
\value{
a dataframe with differential inclusion information
}
\description{
Extract differentially included AS events from HIT Index data
}
\examples{

pdir <- system.file("extdata", package="SpliceImpactR")
dataDirectory <- paste0(pdir, "/rawData/")
test_group <- paste0(dataDirectory, c("test1","test2", "test3"))
control_group <- paste0(dataDirectory, c("control1", "control2", "control3"))

result <- differential_inclusion_rMATS(test_names = test_group,
                                          control_names = control_group,
                                          et = "SE",
                                          outlier_threshold = "Inf",
                                          minReads = 10,
                                          min_prop_samples = 0,
                                          chosen_method = "qbGLM"
                                          )

}
