% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHitCompare.R
\name{diHIT_plots}
\alias{diHIT_plots}
\title{get HITindex plots}
\usage{
diHIT_plots(final_results)
}
\arguments{
\item{final_results}{from getDifferentialHIT}
}
\value{
figures showing hit index plots (dot plot, volcano)
}
\description{
get HITindex plots
}
\keyword{internal}
