% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domains.R
\name{domains_on_exons}
\alias{domains_on_exons}
\title{Internal helper that returns all domain identifiers overlapping a given
set of exons within a transcript.}
\usage{
domains_on_exons(Dexon, tx, exons_vec)
}
\arguments{
\item{Dexon}{Output of [build_domain_lookup()] (`Dexon` element).}

\item{tx}{Character scalar; transcript ID.}

\item{exons_vec}{Character vector of exon IDs.}
}
\value{
Character vector of unique domain identifiers for those exons.
}
\description{
Internal helper that returns all domain identifiers overlapping a given
set of exons within a transcript.
}
\keyword{internal}
