% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domains.R
\name{domains_on_protein}
\alias{domains_on_protein}
\title{Internal helper returning all domains mapped to a given transcript's protein.}
\usage{
domains_on_protein(Dtx, tx)
}
\arguments{
\item{Dtx}{Output of [build_domain_lookup()] (`Dtx` element).}

\item{tx}{Character scalar; transcript ID.}
}
\value{
Character vector of domain identifiers.
}
\description{
Internal helper returning all domains mapped to a given transcript's protein.
}
\keyword{internal}
