% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domain_enrichment.R
\name{.db_remove_domain}
\alias{.db_remove_domain}
\title{Removes the part of a domain identifier after the first semicolon.}
\usage{
.db_remove_domain(id)
}
\arguments{
\item{id}{Character vector of domain identifiers (e.g. `"Pfam;Kinase"`).}
}
\value{
Character vector containing only the database name (e.g. `"Pfam"`).
}
\description{
Removes the part of a domain identifier after the first semicolon.
}
\keyword{internal}
