% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domain_enrichment.R
\name{.domain_remove_db}
\alias{.domain_remove_db}
\title{Removes the database portion from a semicolon-delimited domain ID.}
\usage{
.domain_remove_db(id)
}
\arguments{
\item{id}{Character vector of domain identifiers (e.g. `"Pfam;Kinase"`).}
}
\value{
Character vector of domain names (e.g. `"Kinase"`).
}
\description{
Removes the database portion from a semicolon-delimited domain ID.
}
\keyword{internal}
