% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_hit_index.R
\name{.find_hitindex_files}
\alias{.find_hitindex_files}
\title{Locate HIT index PSI files within a directory or return the file if directly provided.}
\usage{
.find_hitindex_files(p)
}
\arguments{
\item{p}{Directory path or file path.}
}
\value{
Character vector of full paths to matching files.
}
\description{
Locate HIT index PSI files within a directory or return the file if directly provided.
}
\keyword{internal}
