% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domain_enrichment.R
\name{.parse_listcol}
\alias{.parse_listcol}
\title{Internal helper that normalizes list columns or delimited strings
of domain identifiers into unique trimmed character vectors.}
\usage{
.parse_listcol(x, delim = "[,;|[:space:]]+")
}
\arguments{
\item{x}{List or character vector of domain entries.}

\item{delim}{Regular expression used to split delimited strings.}
}
\value{
A list of character vectors, one per input element.
}
\description{
Internal helper that normalizes list columns or delimited strings
of domain identifiers into unique trimmed character vectors.
}
\keyword{internal}
