% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_hit_index.R
\name{.read_exon_files}
\alias{.read_exon_files}
\title{Read exon-level counts or annotations used in HIT index files.}
\usage{
.read_exon_files(path, columns = c("gene", "exon", "ID"))
}
\arguments{
\item{path}{Base path to HIT index outputs (excluding suffix like ".exon").}

\item{columns}{Columns to select from the exon file.}
}
\value{
A data.table with the specified columns.
}
\description{
Read exon-level counts or annotations used in HIT index files.
}
\keyword{internal}
