% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_hit_index.R
\name{.read_one_hit}
\alias{.read_one_hit}
\title{Read and merge a single HIT index PSI file with its exon annotations.}
\usage{
.read_one_hit(f, sample, condition)
}
\arguments{
\item{f}{Path to a .AFEPSI/.ALEPSI/etc. file.}

\item{sample}{Sample name.}

\item{condition}{Experimental condition.}
}
\value{
A standardized data.table with inclusion/exclusion metrics and metadata.
}
\description{
Read and merge a single HIT index PSI file with its exon annotations.
}
\keyword{internal}
