% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{.si_bfc_get_rds}
\alias{.si_bfc_get_rds}
\title{Load cached local RDS from BiocFileCache (internal)}
\usage{
.si_bfc_get_rds(bfc, rname)
}
\arguments{
\item{bfc}{A `BiocFileCache` object.}

\item{rname}{Cache key name.}
}
\value{
Cached R object or `NULL` if not present.
}
\description{
Load cached local RDS from BiocFileCache (internal)
}
\keyword{internal}
