% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{.si_bfc_put_rds}
\alias{.si_bfc_put_rds}
\title{Store R object in BiocFileCache as local RDS (internal)}
\usage{
.si_bfc_put_rds(bfc, rname, obj)
}
\arguments{
\item{bfc}{A `BiocFileCache` object.}

\item{rname}{Cache key name.}

\item{obj}{R object to serialize.}
}
\value{
Invisible cached file path.
}
\description{
Store R object in BiocFileCache as local RDS (internal)
}
\keyword{internal}
