% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{.si_get_annotation_mode_guide}
\alias{.si_get_annotation_mode_guide}
\title{Build short get_annotation mode/use-case guide (internal)}
\usage{
.si_get_annotation_mode_guide()
}
\value{
Character scalar suitable for appending to error messages.
}
\description{
Build short get_annotation mode/use-case guide (internal)
}
\keyword{internal}
