% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_features.R
\name{.si_use_ensembl_mart}
\alias{.si_use_ensembl_mart}
\title{Connect to Ensembl BioMart for a release (internal)}
\usage{
.si_use_ensembl_mart(
  dataset,
  version,
  biomart = "genes",
  host = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dataset}{Ensembl dataset (e.g. `"hsapiens_gene_ensembl"`).}

\item{version}{Ensembl release; its e<release>.ensembl.org host is used
unless `host` is given.}

\item{biomart}{Biomart name, default `"genes"`.}

\item{host}{Optional BioMart host, used as given.}

\item{verbose}{Logical passed to `biomaRt::useEnsembl()`.}
}
\value{
A `Mart` object.
}
\description{
Connect to Ensembl BioMart for a release (internal)
}
\keyword{internal}
