% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domain_enrichment.R
\name{enrich_by_db}
\alias{enrich_by_db}
\title{Run domain enrichment by database}
\usage{
enrich_by_db(hits, background, dbs, ...)
}
\arguments{
\item{hits, background}{See [enrich_domains_hypergeo()].}

\item{dbs}{Character vector of database prefixes to test.}

\item{...}{Additional arguments passed to [enrich_domains_hypergeo()].}
}
\value{
Combined `data.table` with an added `database` column.
}
\description{
Convenience wrapper for [enrich_domains_hypergeo()] that runs the
enrichment test separately for each database (e.g. `"Pfam"`,
`"SMART"`) and combines the results.
}
\examples{

ex <- load_example_data("sample_frame")
sample_frame <- ex$sample_frame
hit_index <- get_hitindex(sample_frame)
res <- get_differential_inclusion(hit_index)
annotation_df <- load_example_data("annotation_df")$annotation_df
matched <- get_matched_events_chunked(res, annotation_df$annotations, chunk_size = 2000)
x_seq <- attach_sequences(matched, annotation_df$sequences)
pairs <- get_pairs(x_seq, source="multi")
seq_compare <-compare_sequence_frame(pairs, annotation_df$annotations)
annotation_df <- get_annotation(load = 'test')
interpro_features <- get_protein_features(c("interpro"), annotation_df$annotations, timeout = 600, test = TRUE)
protein_feature_total <- get_comprehensive_annotations(list(interpro_features))

exon_features <- get_exon_features(annotation_df$annotations, protein_feature_total)

hits_domain <- get_domains(seq_compare, exon_features)

bg <- get_background(source = "annotated",
                     annotations = annotation_df$annotations,
                     protein_features = protein_feature_total)

enriched_domains <- enrich_by_db(hits_domain, bg, dbs = 'interpro')
print(enriched_domains)


}
\seealso{
[enrich_by_event()]
}
