% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameShiftDetector.R
\name{findContinuousIndels}
\alias{findContinuousIndels}
\title{Idenfity continuous indels}
\usage{
findContinuousIndels(indels)
}
\arguments{
\item{indels}{the indels from an alignment}
}
\value{
continuous indels
}
\description{
Idenfity continuous indels
}
