% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameShiftDetector.R
\name{frameShiftDetectorSum}
\alias{frameShiftDetectorSum}
\title{Return both shift and specific results for a given alignment}
\usage{
frameShiftDetectorSum(df, i)
}
\value{
both specific and shift results
}
\description{
Return both shift and specific results for a given alignment
}
