% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameShiftDetector.R
\name{fsDirectShift}
\alias{fsDirectShift}
\title{Align and classify 2 sequences using shift method,}
\usage{
fsDirectShift(seq1, seq2)
}
\value{
both specific and shift results
}
\description{
Align and classify 2 sequences using shift method,
}
