% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullASoutcome.R
\name{fullASoutcome}
\alias{fullASoutcome}
\title{Extract differentially included AFE, ALE from HIT Index data}
\usage{
fullASoutcome(
  as_types = c("AFE", "ALE", "HFE", "HLE", "SE", "MXE", "RI", "A5SS", "A3SS"),
  output_directory,
  data_directory,
  data_df,
  outlier_handle = "Inf",
  cutoff = 0.1,
  cores = 1,
  bg_pre = NA,
  tti_location = "/projectnb/evolution/zwakefield/allison_mettl/analysis/sir/",
  mOverlap = 0.05,
  s_gtf,
  plotAlignments = FALSE,
  transcripts,
  translations,
  biomart_data,
  max_zero_prop = 0.5,
  min_prop_samples = 0.5,
  hitCompare = T,
  chosen_method = "nbGLM"
)
}
\arguments{
\item{as_types}{a vector of the type of alternative splicing events that are being investigated}

\item{output_directory}{the path (ending with '/') that the output is desired}

\item{data_directory}{the path to the data, in the format specified by organizeSamples}

\item{outlier_handle}{value to threshold the outlier detection}

\item{cutoff}{diInclusion cutoff to use to identify significance}

\item{bg_pre}{if bg was made earlier, param to give premade bg}

\item{tti_location}{location of previously made transcript-transcript interactions network}

\item{mOverlap}{minimum overlap to call an exon as matched to annotation}

\item{plotAlignments}{whether to plot alignments from getPaired/matchAlignType}

\item{transcripts}{from getTranscripts}

\item{translations}{from getTranslations}

\item{max_zero_prop}{max prop of samples that can be 0}

\item{min_prop_samples}{min prop of samples from each phenotype required to show a specific event}

\item{gtf}{output from setup_gtf}
}
\value{
nothing in R, output to the output_directory
}
\description{
Extract differentially included AFE, ALE from HIT Index data
}
