% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_c_nucs}
\alias{get_c_nucs}
\title{setting up transcriptions}
\usage{
get_c_nucs(transcripts_location)
}
\arguments{
\item{transcripts_location}{location of transcripts file}
}
\description{
setting up transcriptions
}
