% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_c_trans}
\alias{get_c_trans}
\title{setting up translations}
\usage{
get_c_trans(translations_location)
}
\arguments{
\item{translations_location}{location of translations file}
}
\description{
setting up translations
}
