% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_gtf}
\alias{get_gtf}
\title{wrapper for setup_gtf}
\usage{
get_gtf(location, cores)
}
\arguments{
\item{location}{location of gtf file}

\item{cores}{number of cores requested}
}
\value{
annotated gtf dataframe
}
\description{
wrapper for setup_gtf
}
