% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{getAnnotation}
\alias{getAnnotation}
\title{wrapper for setup_gtf}
\usage{
getAnnotation(biomart_data)
}
\arguments{
\item{biomart_data}{output from setup_biomart}
}
\value{
annotated gtf and others from setup_gtf
}
\description{
wrapper for setup_gtf
}
