% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBackground.R
\name{getBackground}
\alias{getBackground}
\title{getBackground set of transcripts in samples}
\usage{
getBackground(input, mOverlap, cores, exon_type, pdir, output_location, gtf)
}
\arguments{
\item{input}{the path containing the .exon file}

\item{mOverlap}{overlap to identify a match to annotation}

\item{exon_type}{placeholder for other functions}

\item{pdir}{location of the package}

\item{output_location}{location to make background directory}

\item{gtf}{gtf dataframe from setup_gtf}
}
\value{
matched : matched transcripts dataframe, bed : bed file of the matched transcripts
proBed : output for further functions with protein code and protein info,
proFast : fasta file of proteins identified in proBed
}
\description{
getBackground set of transcripts in samples
}
\keyword{internal}
