% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHitCompare.R
\name{getDifferentialHIT}
\alias{getDifferentialHIT}
\title{get differential HIT}
\usage{
getDifferentialHIT(merged_data)
}
\arguments{
\item{merged_data}{data from internal working of getHITCompare}
}
\value{
final_results a dataframe for HIT Index summary
}
\description{
get differential HIT
}
\keyword{internal}
