% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDomainData.R
\name{getDomainData}
\alias{getDomainData}
\title{idenitify enriched domains across phenotype}
\usage{
getDomainData(
  fg,
  bg,
  pfg,
  pfam,
  cores = 1,
  output_location,
  fdr_use = 0.05,
  min_sample_success = 3,
  engine = c("FunFam", "Gene3D", "CDD", "PANTHER", "SMART", "ProSiteProfiles", "Pfam",
    "SUPERFAMILY", "MobiDBLite", "Coils", "PRINTS", "ProSitePatterns", "PIRSF",
    "NCBIfam", "Hamap")[7],
  repeatingDomains = FALSE,
  topViz = 15
)
}
\arguments{
\item{fg}{foreground from getForeground}

\item{bg}{background from getBackground}

\item{pfam}{pfam from getPfam}

\item{cores}{the number of cores requested}

\item{output_location}{location to make background directory}

\item{fdr_use}{the fdr to set as a threshold}

\item{min_sample_success}{the number of appearances of a domain in the sample set to visualize}

\item{engine}{pfam only for now}

\item{repeatingDomains}{whether to identify repeating domains being enriched or not}

\item{topViz}{the max number of domains to put in each visualization}
}
\value{
the data and the enrichment plots
}
\description{
idenitify enriched domains across phenotype
}
