% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTTI.R
\name{getEnrichmentTTI}
\alias{getEnrichmentTTI}
\title{enrichment helper function}
\usage{
getEnrichmentTTI(
  current_transcript,
  t_impacts,
  fdr,
  transGeneProt,
  backgroundGenes,
  steps,
  output_location
)
}
\value{
geneset enrichment from hypeR
}
\description{
enrichment helper function
}
\keyword{internal}
