% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameShiftDetector.R
\name{getFLOverlap}
\alias{getFLOverlap}
\title{Get 1st, last, or general overlapping exons between 2 transcripts}
\usage{
getFLOverlap(transcript1, transcript2, ex, coding_exonsX, eld)
}
\arguments{
\item{transcript1}{transcript 1}

\item{transcript2}{transcript 2}

\item{ex}{exon_type}

\item{coding_exonsX}{coding exons from initFrameShift}

\item{eld}{exon_length_df from initFrameShift}
}
\value{
a character vector of types of alignmetns of each pair
}
\description{
Get 1st, last, or general overlapping exons between 2 transcripts
}
\keyword{internal}
