% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getForeground.R
\name{getForeground}
\alias{getForeground}
\title{get Foreground differentially included transcripts in samples}
\usage{
getForeground(
  input,
  test_names,
  control_names,
  thresh = 0.1,
  fdr = 0.05,
  mOverlap,
  exon_type,
  pdir,
  output_location,
  cores = 1,
  gtf,
  max_zero_prop = 0.5,
  min_prop_samples = 0.5
)
}
\arguments{
\item{input}{diExon info from the differentially included functions}

\item{thresh}{delta psi threshold to filter}

\item{fdr}{adj p to filter for}

\item{mOverlap}{overlap to identify a match to annotation}

\item{exon_type}{placeholder for other functions}

\item{pdir}{location of the package}

\item{output_location}{location to make background directory}

\item{gtf}{full output from setup_gtf}

\item{max_zero_prop}{max number of zeros in to count an exon}

\item{min_prop_samples}{min proportion of samples needed to not be counted as outliers}
}
\value{
matched : matched transcripts dataframe, bed : bed file of the matched transcripts
proBed : output for further functions with protein code and protein info,
proFast : fasta file of proteins identified in proBed
}
\description{
get Foreground differentially included transcripts in samples
}
