% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameShiftDetector.R
\name{getFrameShift}
\alias{getFrameShift}
\title{identify frame shifts due to aRNAp events}
\usage{
getFrameShift(fC, et, newgtf, exon_data)
}
\arguments{
\item{fC}{addinf col}

\item{et}{aRNAp type}

\item{newgtf}{newgtf from setupGTF}

\item{exon_data}{exon data data frame}
}
\value{
a character vector of types of alignmetns of each pair
}
\description{
identify frame shifts due to aRNAp events
}
\keyword{internal}
