% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameShiftDetector.R
\name{getFrameShiftInit}
\alias{getFrameShiftInit}
\title{initialize get frame shift with data}
\usage{
getFrameShiftInit(newgtf, exon_data)
}
\arguments{
\item{newgtf}{from setupgtf}

\item{exon_data}{exon_data from earlier}
}
\value{
multiple dfs with different info and organization related to coding exons
}
\description{
initialize get frame shift with data
}
\keyword{internal}
