% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfxnlASoutcome.R
\name{getfxnlASoutcome}
\alias{getfxnlASoutcome}
\title{Full run of the pipeline for a single alternative splicing event type}
\usage{
getfxnlASoutcome(
  output_location,
  test_group,
  control_group,
  data_df,
  exon_type,
  cutoff = 0.1,
  outlier_handle = "4/n",
  cores = 1,
  tti_location = "",
  full_pipe = TRUE,
  bg = NA,
  mOverlap = 0.05,
  gtf,
  plotAlignments = FALSE,
  transcripts,
  translations,
  biomart_data,
  max_zero_prop = 0.5,
  min_prop_samples = 0.5
)
}
\arguments{
\item{output_location}{location to make  directory}

\item{test_group}{the paths of one phenotype}

\item{control_group}{the paths of one phenotype}

\item{data_df}{dataframe with sample paths, type, and phenotype}

\item{exon_type}{type of exon being investigated}

\item{cutoff}{cutoff for significance of differential inclusion}

\item{tti_location}{location of previuously made tti or ""}

\item{full_pipe}{if TRUE, doesn't save output to R, only writes to output_location}

\item{bg}{whether bg preran or needs making}

\item{mOverlap}{overlap to identify a match to annotation}

\item{gtf}{output from setup_gtf}

\item{plotAlignments}{whether to output paired alignments}

\item{transcripts}{from getTranscripts}

\item{translations}{from getTranslations}

\item{max_zero_prop}{max prop of samples that can be 0}

\item{min_prop_samples}{min prop of samples from each phenotype required to show a specific event}
}
\value{
nothing or all output from pipeline
}
\description{
Full run of the pipeline for a single alternative splicing event type
}
