% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntegrative.R
\name{getIntegratedResults}
\alias{getIntegratedResults}
\title{Perform various integrative analyses on multiple aRNAp results}
\usage{
getIntegratedResults(fg_list, pfg_list, domain_list)
}
\arguments{
\item{fg_list}{a list of the fglfc.csv output from getForeground, names are the event types}

\item{pfg_list}{a list of the paired_output from getPaired, names are the event types}

\item{domain_list}{a list of the data output from getDomainData, names are the event types}
}
\value{
a dataframe with summary stats for each event type supplied along with various plots
}
\description{
Perform various integrative analyses on multiple aRNAp results
}
