% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLengthComparison.R
\name{getLengthComparison}
\alias{getLengthComparison}
\title{Get comparison between length across isoform swaps due to AS}
\usage{
getLengthComparison(paired_df, output_location)
}
\arguments{
\item{paired_df}{output df from getPaired}

\item{output_location}{location where everything is being saved}
}
\value{
1 combined plot.
}
\description{
Get comparison between length across isoform swaps due to AS
}
\examples{
getOverviewComparison(paired_combined_rows, output_path)
}
