% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{getMXE_internal}
\alias{getMXE_internal}
\title{helper for MXEmatcher}
\usage{
getMXE_internal(g, indices, clusion, strVar, minOverlap)
}
\value{
rownums for the clusion
}
\description{
helper for MXEmatcher
}
\keyword{internal}
