% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPaired.R
\name{getPaired}
\alias{getPaired}
\title{get paired as events / exons}
\usage{
getPaired(
  foreground,
  et,
  nucleotides,
  newGTF,
  cores = 1,
  output_location,
  saveAlignments = T
)
}
\arguments{
\item{foreground}{proBed from getForeground}

\item{et}{exon_type being investigated}

\item{nucleotides}{made from setup get_c_nucs}

\item{newGTF}{from setup get_gtf}

\item{cores}{number of requested cores}

\item{output_location}{location to make background directory}

\item{saveAlignments}{bool to save alignment pdfs}

\item{pdir}{location of the package}
}
\value{
figures and dataframes with paired data
}
\description{
get paired as events / exons
}
