% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProximalShift.R
\name{getProximalShift}
\alias{getProximalShift}
\title{get paired as events / exons}
\usage{
getProximalShift(type, exon_pairs, ep_supp, output_location)
}
\arguments{
\item{type}{event type}

\item{exon_pairs}{recent dataframe made from getPaired}

\item{ep_supp}{recent dataframe made from getPaired}

\item{output_location}{location for saving}
}
\value{
figure showing proximal / distal shift and table of results
}
\description{
get paired as events / exons
}
