% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameShiftDetector.R
\name{getRescue}
\alias{getRescue}
\title{check for rescue of the frame shifts}
\usage{
getRescue(transcript1, transcript2, e1, e2, eld, filterDownstream = FALSE)
}
\arguments{
\item{transcript1}{first transcript}

\item{transcript2}{second transcript}

\item{e1}{first exon}

\item{e2}{second exon}

\item{eld}{exon_length_df from previous function}

\item{filterDownstream}{param to choose whether or not to filter to downstream of exons}
}
\value{
either noRescues or the transcript and exons that contain rescue
}
\description{
check for rescue of the frame shifts
}
\keyword{internal}
