% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{getRI_internal}
\alias{getRI_internal}
\title{helper for RImatcher}
\usage{
getRI_internal(g, indices, clusion, strVar, minOverlap)
}
\value{
rownums for the clusion
}
\description{
helper for RImatcher
}
\keyword{internal}
