% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_models.R
\name{getSizeFactors}
\alias{getSizeFactors}
\title{Extract size factors for each sample}
\usage{
getSizeFactors(df)
}
\arguments{
\item{df}{dataframe with key info for size factor calc}
}
\value{
a dataframe with sizeFactors added
}
\description{
Extract size factors for each sample
}
\keyword{internal}
