% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTranscript.R
\name{getTranscriptBackground}
\alias{getTranscriptBackground}
\title{gets the transcripts to match to each as event or exon}
\usage{
getTranscriptBackground(gtf, redExon, ex_type, minOverlap = 0.05, cores = 1)
}
\arguments{
\item{gtf}{gtf from get_gtf from setup}

\item{redExon}{input of simplified exon coords internally}

\item{minOverlap}{minimum overlap to classify as matched to annotation}

\item{cores}{number of requested cores}

\item{exon_type}{type of exon being queried}
}
\value{
figures and dataframes with paired data
}
\description{
gets the transcripts to match to each as event or exon
}
