% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{getTranslations}
\alias{getTranslations}
\title{setting up translations}
\usage{
getTranslations(translations_location)
}
\arguments{
\item{translations_location}{location of translations file}
}
\value{
modified protein code fasta
}
\description{
setting up translations
}
\examples{

pdir <- system.file("extdata", package = "SpliceImpactR")
dataDirectory <- paste0(pdir, "/")
translations <- getTranslations(translations_location = dataDirectory)

}
