% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTTI.R
\name{getTTI}
\alias{getTTI}
\title{gets the tti interactions with various helper functions to do so}
\usage{
getTTI(
  paired_foreground,
  background,
  pdir = pdir,
  steps = 1,
  max_vertices_for_viz = 5000,
  fdr = 0.05,
  plot_bool = T,
  ppidm_class = c("Gold", "Silver", "Bronze")[1],
  write_igraphs = T,
  output_location = output_location,
  tti_location = output_location
)
}
\arguments{
\item{paired_foreground}{from getPaired}

\item{background}{proBed from getBackground}

\item{pdir}{the directory of the package}

\item{steps}{the number of steps for viz}

\item{max_vertices_for_viz}{max number of vertices to plot, saves time and space}

\item{fdr}{fdr to cut off for geneset enrichment}

\item{plot_bool}{bool to make output plots of not}

\item{ppidm_class}{threshold of ppidm sets to use}

\item{output_location}{location to make output}

\item{tti_location}{if init_tti already performed, location of output or ""}

\item{write_graphs}{save full graph interactions for each with signfiicant changes}

\item{minOverlap}{minimum overlap to classify as matched to annotation}

\item{cores}{number of requested cores}
}
\value{
differences between each tti pair and the overall results
}
\description{
gets the tti interactions with various helper functions to do so
}
