% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTTI.R
\name{getTTIiGraphPlot}
\alias{getTTIiGraphPlot}
\title{graph helper function}
\usage{
getTTIiGraphPlot(
  paired_transcript,
  gene,
  steps,
  full_graph,
  max_vertices_for_viz,
  output_location
)
}
\value{
tti igraphs
}
\description{
graph helper function
}
\keyword{internal}
