% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{HITmatcher}
\alias{HITmatcher}
\title{specific matcher for HIT}
\usage{
HITmatcher(
  i,
  redExon = redExon,
  gtf_filtered = gtf_filtered,
  minOverlap = minOverlap,
  protein_coding_transcripts = protein_coding_transcripts
)
}
\description{
specific matcher for HIT
}
