% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTTI.R
\name{init_ddi}
\alias{init_ddi}
\title{initiates tti network for entire genome, can be very time consumng}
\usage{
init_ddi(
  pdir,
  output_location,
  ppidm_class = c("Gold_Standard", "Gold", "Silver", "Bronze")[1],
  removeDups = TRUE,
  cores = 1,
  pfam_data
)
}
\arguments{
\item{pdir}{the directory of the package}

\item{output_location}{location to make output}

\item{ppidm_class}{threshold of ppidm sets to use}

\item{removeDups}{takes a longer time to init, but faster down the line and saves a lot of memory to remove duplcate vertices}

\item{cores}{number of requested cores}
}
\value{
overall tti network
}
\description{
initiates tti network for entire genome, can be very time consumng
}
