% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPfam.R
\name{initPFAM}
\alias{initPFAM}
\title{get the pfam domains and ip conversions}
\usage{
initPFAM(biomart_data)
}
\value{
exon and transcript level domains
}
\description{
get the pfam domains and ip conversions
}
\keyword{internal}
