% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameShiftDetector.R
\name{irRead}
\alias{irRead}
\title{Get alignment status of introns}
\usage{
irRead(addInf, coding_exons, exon_data, exon_length_df)
}
\arguments{
\item{addInf}{addInf column passed -- extra info}

\item{coding_exons}{coding exons from initFrameShift}

\item{exon_length_df}{exon_length_df from initFrameShift}

\item{exon_length}{exon_data from initFrameShift}
}
\value{
a character vector of types of alignmetns of each pair
}
\description{
Get alignment status of introns
}
\keyword{internal}
