% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dPsiPlot.R
\name{make_dPsiPlot}
\alias{make_dPsiPlot}
\title{gets the tti interactions with various helper functions to do so}
\usage{
make_dPsiPlot(dpsi_df, gtf, thresh = 0.1)
}
\arguments{
\item{dpsi_df}{pre filtering differential inclusion analysis df}

\item{gtf}{gtf from setupGTF}

\item{thresh}{threshold for differential inclusion}
}
\value{
differences between each tti pair and the overall results
}
\description{
gets the tti interactions with various helper functions to do so
}
\keyword{internal}
