% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppi.R
\name{mark_changing_partners_split}
\alias{mark_changing_partners_split}
\title{Helper to annotate PPI changes from DDI and DMI}
\usage{
mark_changing_partners_split(
  ppi,
  gene_id,
  changed_pfam_inc,
  changed_pfam_exc,
  changed_motif_inc = character(),
  changed_motif_exc = character()
)
}
\description{
Helper to annotate PPI changes from DDI and DMI
}
\keyword{internal}
