% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchAlignType.R
\name{matchAlignType}
\alias{matchAlignType}
\title{getPaired helper function to identify align classifications}
\usage{
matchAlignType(
  proBed,
  protCode,
  nucleotides,
  output_location = NULL,
  saveAlignments = FALSE,
  exon_type,
  newgtf,
  exon_data
)
}
\arguments{
\item{proBed}{proBed from getForeground in getPaired}

\item{protCode}{protein code for each protein / transcript / exon associated}

\item{nucleotides}{get_c_nucs from setup}

\item{output_location}{output location}

\item{saveAlignments}{whether to save alignment output pdf}
}
\value{
altered probed and alignment info
}
\description{
getPaired helper function to identify align classifications
}
\keyword{internal}
