% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{matcher}
\alias{matcher}
\title{matches the given locations to annotations and transcripts}
\usage{
matcher(
  ex_type,
  background = FALSE,
  cores = 1,
  redExon_globe,
  minOverlap = 0.05
)
}
\arguments{
\item{ex_type}{type of exon being queried}

\item{background}{whether making background or not}

\item{cores}{number of requested cores}

\item{redExon_globe}{input of simplified exon coords internally}

\item{minOverlap}{minimum overlap to classify as matched to annotation}
}
\value{
figures and dataframes with paired data
}
\description{
matches the given locations to annotations and transcripts
}
