% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{MXmatcher}
\alias{MXmatcher}
\title{specific matcher for MXE}
\usage{
MXmatcher(
  i,
  below_thresh = 0,
  redExon,
  minOverlap = 0.05,
  gtf_transcripts,
  gtf_exons,
  protein_coding_transcripts,
  transcript_starts
)
}
\value{
matched transcript rownumber
}
\description{
specific matcher for MXE
}
