% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_models.R
\name{nbGLM_model}
\alias{nbGLM_model}
\title{Extract significance stats for each aRNAp}
\usage{
nbGLM_model(exon_data, gene_data, exon_of_interest, threshold = Inf)
}
\arguments{
\item{threshold}{threshold for cooks.d}

\item{data}{dataframe with key info for diff inc calc}
}
\value{
p vals, LR vals, cooks d vals
}
\description{
Extract significance stats for each aRNAp
}
\keyword{internal}
