% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeSamples.R
\name{organizeSamples}
\alias{organizeSamples}
\title{This function organizes sample data by creating directories for each sample and copying specific files into them.
It is designed to work within a specific directory structure that includes 'rmats' and 'hit' folders.
## if in structure:
sample
# rmats
# hit-index}
\usage{
organizeSamples(samples, output_location, cores = 1)
}
\description{
This function organizes sample data by creating directories for each sample and copying specific files into them.
It is designed to work within a specific directory structure that includes 'rmats' and 'hit' folders.
## if in structure:
sample
# rmats
# hit-index
}
