% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameShiftDetector.R
\name{overlap}
\alias{overlap}
\title{Find overlap between two sequences}
\usage{
overlap(start1, end1, start2, end2)
}
\arguments{
\item{start1}{sequence 1 start}

\item{end1}{sequence 1 end}

\item{start2}{sequence 2 start}

\item{end2}{sequence 2 end}
}
\value{
which sequences are overlapping
}
\description{
Find overlap between two sequences
}
\keyword{internal}
