% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_inclusion_rMATS.R
\name{paired_rMATS_helper}
\alias{paired_rMATS_helper}
\title{Helper for rMATS pairing operation}
\usage{
paired_rMATS_helper(df, type)
}
\value{
modified data frame for rMATS analysis
}
\description{
Helper for rMATS pairing operation
}
\keyword{internal}
