% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_differential_inclusion.R
\name{plot_di_volcano_dt}
\alias{plot_di_volcano_dt}
\title{Volcano plot for differential inclusion results}
\usage{
plot_di_volcano_dt(di, padj_thr = 0.05, dpsi_thr = 0.1)
}
\arguments{
\item{di}{A `data.frame` or `data.table` containing at least
`delta_psi` and `padj` columns (optionally `event_type`).}

\item{padj_thr}{Numeric. Adjusted p-value threshold (default `0.05`).}

\item{dpsi_thr}{Numeric. Absolute deltaPSI threshold (default `0.1`).}
}
\value{
A `ggplot2` object showing differential inclusion significance.
}
\description{
Generates a volcano plot of deltaPSI vs. -log10(FDR) highlighting significant events.
}
\details{
Significant sites are colored in `deeppink4`; nonsignificant sites are
shown in light grey. Dashed and dotted lines indicate deltaPSI and FDR thresholds.
}
\examples{
ex <- load_example_data("sample_frame")
sample_frame <- ex$sample_frame
hit_index <- get_hitindex(sample_frame)
res <- get_differential_inclusion(hit_index)
plot_di_volcano_dt(res)

}
