% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_models.R
\name{qbGLM_model}
\alias{qbGLM_model}
\title{Extract significance stats for each aRNAp}
\usage{
qbGLM_model(data, threshold = Inf)
}
\arguments{
\item{data}{dataframe with key info for diff inc calc}

\item{threshold}{threshold for cooks.d}
}
\value{
p vals, LR vals, cooks d vals
}
\description{
Extract significance stats for each aRNAp
}
\keyword{internal}
