% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diHelpers.R
\name{qualityFilter}
\alias{qualityFilter}
\title{filter using minimums for identifying events in control/test and max proportion of zeros allowed}
\usage{
qualityFilter(df, nT, nC)
}
\arguments{
\item{nT}{number of test samples}

\item{nC}{number of control samples}

\item{final_data}{from diHIT or diRMATS}
}
\value{
a filtered di inclusion dataframe
}
\description{
filter using minimums for identifying events in control/test and max proportion of zeros allowed
}
\examples{
qualityFilter(diHIT_output)
}
