% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{SEmatcher}
\alias{SEmatcher}
\title{specific matcher for SE}
\usage{
SEmatcher(
  i,
  below_thresh = 0.2,
  redExon,
  minOverlap = 0.05,
  gtf_transcripts,
  gtf_exons,
  protein_coding_transcripts,
  transcript_starts
)
}
\description{
specific matcher for SE
}
