% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gtf.R
\name{setup_gtf}
\alias{setup_gtf}
\title{setting up gtf with exon labels, first, internal, last, hybrid, etc}
\usage{
setup_gtf(gtf_location, cores = 1)
}
\arguments{
\item{gtf_location}{location of gtf file}

\item{cores}{number of cores requested}
}
\description{
setting up gtf with exon labels, first, internal, last, hybrid, etc
}
