% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gtf.R
\name{setupAnnotation}
\alias{setupAnnotation}
\title{setting up gtf with exon labels, first, internal, last, hybrid, etc}
\usage{
setupAnnotation(biomart_data)
}
\value{
annotated gtf and various hybrid info. also gene to transcript to protein naming dataframe
}
\description{
setting up gtf with exon labels, first, internal, last, hybrid, etc
}
\keyword{internal}
