% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gtf.R
\name{setupBiomart}
\alias{setupBiomart}
\title{import / load + save biomaRt data}
\usage{
setupBiomart(save_location, m = "")
}
\value{
various biomaRt results for use downstream in pipeline
}
\description{
import / load + save biomaRt data
}
\keyword{internal}
