% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diHelpers.R
\name{significanceFilter}
\alias{significanceFilter}
\title{filter using delta psi and fdr}
\usage{
significanceFilter(df, fdr = 0.05, d.psi = 0.1)
}
\arguments{
\item{final_data}{from diHIT or diRMATS (or post filtering)}
}
\value{
a filtered di inclusion dataframe
}
\description{
filter using delta psi and fdr
}
\examples{
significanceFilter(diHIT_output, fdr = .05, d.psi = .1)
}
