% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarized_experiment_utils.R
\name{spliceimpact_hit_colsets}
\alias{spliceimpact_hit_colsets}
\title{List predefined paired-hit column subsets}
\usage{
spliceimpact_hit_colsets()
}
\value{
Named list of predefined column vectors for paired-hit accessors.
}
\description{
List predefined paired-hit column subsets
}
\keyword{internal}
