% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpliceImpactR.R
\docType{_PACKAGE}
\name{SpliceImpactR}
\alias{SpliceImpactR}
\title{SpliceImpactR: An R package to identify functional impacts due to 
alternative RNA processing events}
\description{
Works by taking in processed data from the HIT Index and/or rMATS
and identifying how differentially used alternative RNA processing events 
lead to changes in protein function through various means. Primarily this 
is done through protein similarity, functional protein domain analysis, and 
domain-domain interaction changes. Notably, we both identify alterantive 
RNA processing event 'swaps' across condition and are able to perform 
holistic analyses regarding the impact of different RNA processing events.
}
\author{
Zachary Wakefield
}
