% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_features.R
\name{split_into_bits}
\alias{split_into_bits}
\title{Split protein-coding transcript IDs into bounded BioMart batches (internal)}
\usage{
split_into_bits(gtf_df, max_group_size)
}
\arguments{
\item{gtf_df}{Annotation table containing transcript metadata.}

\item{max_group_size}{Maximum number of transcript IDs in each request.}
}
\value{
List of character vectors of version-free transcript IDs.
}
\description{
Split protein-coding transcript IDs into bounded BioMart batches (internal)
}
\keyword{internal}
