% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_models.R
\name{wilcox_model}
\alias{wilcox_model}
\title{Extract significance stats for each aRNAp}
\usage{
wilcox_model(data)
}
\arguments{
\item{data}{dataframe with key info for diff inc calc}
}
\value{
p vals for each event
}
\description{
Extract significance stats for each aRNAp
}
\keyword{internal}
