% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarized_experiment_utils.R
\name{as_dt_from_s4}
\alias{as_dt_from_s4}
\title{Convert S4 slots back to data.table}
\usage{
as_dt_from_s4(
  x,
  slot = c("raw_events", "di_events", "paired_hits"),
  keep_internal_keys = FALSE
)
}
\arguments{
\item{x}{`SpliceImpactResult`}

\item{slot}{One of `raw_events`, `di_events`, `paired_hits`.}

\item{keep_internal_keys}{Keep internal key columns (`raw_key`, `di_key`, `pair_key`).}
}
\value{
`data.table`
}
\description{
Convert S4 slots back to data.table
}
