% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{ASmatcher}
\alias{ASmatcher}
\title{specific matcher for A5SS/A3SS}
\usage{
ASmatcher(
  i,
  below_thresh = 0.2,
  redExon,
  minOverlap = 0.05,
  gtf_transcripts,
  gtf_exons,
  protein_coding_transcripts,
  transcript_starts,
  whichType
)
}
\value{
matched transcript rownumber
}
\description{
specific matcher for A5SS/A3SS
}
\keyword{internal}
