% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diHelpers.R
\name{diColor}
\alias{diColor}
\title{add colors to differential inclusion analysis for plotting}
\usage{
diColor(de_df, color_thresh = 0.2)
}
\arguments{
\item{de_df}{from differential inclusion functions}

\item{color_thresh}{the threshold to color by, default of .2}
}
\value{
the input dataframe with color info added
}
\description{
add colors to differential inclusion analysis for plotting
}
\keyword{internal}
