% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_inclusion_HITindex.R
\name{differential_inclusion_HITindex}
\alias{differential_inclusion_HITindex}
\title{Extract differentially included AFE, ALE from HIT Index data}
\usage{
differential_inclusion_HITindex(
  test_names,
  control_names,
  et,
  outlier_threshold = c("4/n", "1", "Inf")[3],
  minReads = 10,
  min_prop_samples = 0.5,
  chosen_method = "qbGLM",
  geneIndependent = TRUE
)
}
\arguments{
\item{test_names}{character vector of the path of test_names (eg:
'/path/to/test1', with the sample located '/path/to/test1.AFEPSI')}

\item{control_names}{character vector of the path of control_names (eg:
'/path/to/control1', with the sample located '/path/to/control1.AFEPSI')}

\item{et}{string of the event type (eg: 'AFE')}

\item{outlier_threshold}{the thresholding of the cooks distance,
no outlier removal is "Inf"}

\item{minReads}{threshold to count an exon as present}

\item{min_prop_samples}{minimum proportion of samples in
either phenotype an exon has to be present in to be kept}

\item{chosen_method}{stats model abbreviation}

\item{geneIndependent}{bool to (TRUE) remove non expressed genes}
}
\value{
a dataframe with differential inclusion information
}
\description{
Extract differentially included AFE, ALE from HIT Index data
}
\examples{

pdir <- system.file("extdata", package="SpliceImpactR")
dataDirectory <- paste0(pdir, "/rawData/")
test_group <- paste0(dataDirectory, c("test1","test2", "test3"))
control_group <- paste0(dataDirectory, c("control1", "control2", "control3"))

result <- differential_inclusion_HITindex(test_names = test_group,
                                          control_names = control_group,
                                          et = "AFE",
                                          outlier_threshold = "Inf",
                                          minReads = 10,
                                          min_prop_samples = 0,
                                          chosen_method = "qbGLM"
                                          )

}
