% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_inclusion_rMATS.R
\name{differential_inclusion_rMATS}
\alias{differential_inclusion_rMATS}
\title{Extract differentially included AS events from HIT Index data}
\usage{
differential_inclusion_rMATS(
  control_names,
  test_names,
  et,
  outlier_threshold = c("4/n", "1", "Inf")[3],
  minReads = 10,
  min_prop_samples = 0.5,
  chosen_method
)
}
\arguments{
\item{control_names}{a vector of control_names}

\item{test_names}{a vector of test_names}

\item{et}{string of the exon_type}

\item{outlier_threshold}{the thresholding of the cooks distance, no outlier removal is "Inf"}

\item{min_prop_samples}{min prop of samples to require an event to be identified in}

\item{chosen_method}{method for stats analysis}

\item{cores}{the number of cores requested}
}
\value{
a dataframe with differential inclusion information
}
\description{
Extract differentially included AS events from HIT Index data
}
