% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameShiftDetector.R
\name{fsDirectSpecific}
\alias{fsDirectSpecific}
\title{Align and classify 2 sequences using specific method -- using nucleotides finding indels of non-3 multiples or leads or non-3}
\usage{
fsDirectSpecific(seq1, seq2)
}
\value{
both specific and shift results
}
\description{
Align and classify 2 sequences using specific method -- using nucleotides finding indels of non-3 multiples or leads or non-3
}
