% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullASoutcome.R
\name{fullASoutcome}
\alias{fullASoutcome}
\title{Extract differentially included AFE, ALE from HIT Index data}
\usage{
fullASoutcome(
  as_types = c("AFE", "ALE", "HFE", "HLE", "SE", "MXE", "RI", "A5SS", "A3SS"),
  output_directory,
  data_directory,
  data_df,
  outlier_handle,
  cutoff = 0.1,
  cores = 1,
  bg_pre = NA,
  tti_location = "/projectnb/evolution/zwakefield/allison_mettl/analysis/sir/",
  mOverlap = 0.05,
  gtf
)
}
\arguments{
\item{as_types}{a vector of the type of alternative splicing events that are being investigated}

\item{output_directory}{the path (ending with '/') that the output is desired}

\item{data_directory}{the path to the data, in the format specified by organizeSamples}

\item{outlier_handle}{value to threshold the outlier detection}

\item{cutoff}{diInclusion cutoff to use to identify significance}

\item{bg_pre}{if bg was made earlier, param to give premade bg}

\item{tti_location}{location of previously made transcript-transcript interactions network}

\item{mOverlap}{minimum overlap to call an exon as matched to annotation}

\item{gtf}{gtf data_frame from setup_gtf}
}
\value{
nothing in R, output to the output_directory
}
\description{
Extract differentially included AFE, ALE from HIT Index data
}
