% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multi_analysis.R
\name{get_domain_gene_for_enrichment}
\alias{get_domain_gene_for_enrichment}
\title{Extract Domain-Altering Genes for Enrichment}
\usage{
get_domain_gene_for_enrichment(hits)
}
\arguments{
\item{hits}{Data frame with domain-annotation columns.}
}
\value{
Character vector of gene IDs.
}
\description{
Return genes whose inclusion/exclusion isoforms show unique protein domain gain or loss.
}
\examples{
\donttest{
ex <- load_example_data("sample_frame")
sample_frame <- ex$sample_frame
hit_index <- get_hitindex(sample_frame)
res <- get_differential_inclusion(hit_index)
annotation_df <- load_example_data("annotation_df")$annotation_df
matched <- get_matched_events_chunked(res, annotation_df$annotations, chunk_size = 2000)
x_seq <- attach_sequences(matched, annotation_df$sequences)
pairs <- get_pairs(x_seq, source="multi")
seq_compare <-compare_sequence_frame(pairs, annotation_df$annotations)
annotation_df <- get_annotation(load = 'test')
interpro_features <- get_protein_features(c("interpro"), annotation_df$annotations, timeout = 600, test = TRUE)
protein_feature_total <- get_comprehensive_annotations(list(interpro_features))

exon_features <- get_exon_features(annotation_df$annotations, protein_feature_total)

hits_domain <- get_domains(seq_compare, exon_features)

bg <- get_background(source = "hit_index",
                     input = sample_frame,
                     annotations = annotation_df$annotations,
                     protein_features = protein_feature_total)
enrichment <- get_enrichment(get_domain_gene_for_enrichment(hits_domain), bg$gene_id, species = 'human', 'ensembl', 'MSigDB:H')
print(enrichment)
}
}
