% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{get_example_data}
\alias{get_example_data}
\title{Helper to fetch test files for vignettes / tests}
\usage{
get_example_data(filename)
}
\arguments{
\item{filename}{file to probe}
}
\value{
proper path to example data
}
\description{
Helper to fetch test files for vignettes / tests
}
\keyword{internal}
