% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_hit_index.R
\name{get_hitindex}
\alias{get_hitindex}
\title{Load HIT index PSI files for one or more samples/conditions.}
\usage{
get_hitindex(paths_df, keep_annotated_first_last = FALSE)
}
\arguments{
\item{paths_df}{Data.frame with columns: \code{path}, \code{condition}, and optionally \code{sample_name}.}

\item{keep_annotated_first_last}{Logical; if TRUE, retain only annotated first/last exons and normalize PSI.}
}
\value{
A standardized `data.table` of HIT index PSI values with inclusion/exclusion and metadata.
}
\description{
Load HIT index PSI files for one or more samples/conditions.
}
\examples{
sample_frame <- data.frame(path = c(check_extdata_dir('rawData/control_S5/'),
                                    check_extdata_dir('rawData/control_S6/'),
                                    check_extdata_dir('rawData/control_S7/'),
                                    check_extdata_dir('rawData/control_S8/'),
                                    check_extdata_dir('rawData/case_S1/'),
                                    check_extdata_dir('rawData/case_S2/'),
                                    check_extdata_dir('rawData/case_S3/'),
                                    check_extdata_dir('rawData/case_S4/')),
                           sample_name  = c("S5", "S6", "S7", "S8", "S1", "S2", "S3", "S4"),
                           condition    = c("control", "control", "control", "control", "case",  "case",  "case",  "case"),
                           stringsAsFactors = FALSE)
hit_index <- get_hitindex(sample_frame, keep_annotated_first_last = TRUE)

}
