% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_features.R
\name{get_linear_motifs}
\alias{get_linear_motifs}
\title{Get short linear motif validated instances from ELM and convert to pf form}
\usage{
get_linear_motifs(
  gtf_df,
  protein_seqs,
  species = c("hsapiens_gene_ensembl", "mmusculus_gene_ensembl"),
  release = 109,
  ensembl_mirror = NULL
)
}
\arguments{
\item{gtf_df}{A \code{data.frame} or \code{data.table} containing GTF
annotations; used to derive chromosome groups for batching.}

\item{protein_seqs}{only necessary if loading SLiMs from elm get_annotation() 
(default \code{"sequences"}) output}

\item{species}{Character string giving the Ensembl BioMart
dataset (default \code{"hsapiens_gene_ensembl"}). For mouse, use
\code{"mmusculus_gene_ensembl"}.}

\item{release}{Release version from Ensembl associated with the GENCODE
version used in [get_annotation()]. See the GENCODE human release listing
to map GENCODE and Ensembl versions.}

\item{ensembl_mirror}{Optional Ensembl mirror passed to the BioMart
connector.}
}
\value{
A \code{data.table} containing protein feature annotations
  including transcript and peptide IDs, feature start/end positions,
  and database-specific identifiers (e.g., InterPro accession) for elm SLiMs
}
\description{
Get short linear motif validated instances from ELM and convert to pf form
}
\details{
Here we access ELM's SLiM database to pull instances and classes and use
BiomaRt to match up uniprot to ensembl + confirm with regex checks
}
\keyword{internal}
