% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_features.R
\name{get_manual_features}
\alias{get_manual_features}
\title{Incorporate user-supplied protein features}
\usage{
get_manual_features(
  manual_features,
  gtf_df,
  biomaRt_features = NULL,
  load_path = NULL,
  save_path = NULL
)
}
\arguments{
\item{manual_features}{Data.frame or data.table with at least
\code{name}, \code{start}, \code{stop} amino acid, and one of
\code{ensembl_transcript_id} or \code{ensembl_peptide_id}.}

\item{gtf_df}{get_annotation annotation output}

\item{biomaRt_features}{Optional \code{data.table} of features from
[get_protein_features()] to merge with.}

\item{load_path}{Optional path to load precomputed manual features.}

\item{save_path}{Optional path to save the combined feature table.}
}
\value{
A \code{data.table} of manual (and optionally combined)
  protein features.
}
\description{
Converts a manual feature table into the standardized long format
and optionally merges it with biomaRt-derived features.
}
\examples{
annotation_df <- load_example_data("annotation_df")$annotation_df
user_df <- data.frame(
  ensembl_transcript_id = c(
    "ENST00000511072","ENST00000374900","ENST00000373020","ENST00000456328",
    "ENST00000367770","ENST00000331789","ENST00000335137","ENST00000361567",
    NA,                    "ENST00000380152"
  ),
  ensembl_peptide_id = c(
    "ENSP00000426975", NA,                   "ENSP00000362048","ENSP00000407743",
    "ENSP00000356802","ENSP00000326734", NA,                  "ENSP00000354587",
    "ENSP00000364035", NA
  ),
  name = c(
    "Low complexity","Transmembrane helix","Coiled-coil","Signal peptide",
    "Transmembrane helix","Low complexity","Coiled-coil","Transmembrane helix",
    "Signal peptide","Low complexity"
  ),
  start = c(80L, 201L, 35L, 1L, 410L, 150L, 220L, 30L, 1L, 300L),
  stop  = c(120L,223L, 80L, 20L, 430L, 190L, 260L, 55L, 24L, 360L),
  database   = c("seg","tmhmm","ncoils","signalp","tmhmm","seg","ncoils","tmhmm","signalp", NA),
  alt_name   = c(NA,"TMhelix",NA,"SignalP-noTM", "TMhelix", NA, NA, "TMhelix", "SignalP-TAT", NA),
  feature_id = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA)
)
user_features <- get_manual_features(user_df, annotation_df$annotations)
print(user_features)

}
\seealso{
[add_user_features()]
}
