% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppi.R
\name{get_ppi_interactions}
\alias{get_ppi_interactions}
\title{Pull PPI from SpliceImpactR's data}
\usage{
get_ppi_interactions()
}
\description{
Generation details in inst/scripts
}
\examples{
ppi_int <- get_ppi_interactions()
print(ppi_int)

}
