% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppi.R
\name{get_ppi_switches}
\alias{get_ppi_switches}
\title{Compute PPI gains and losses for inclusion/exclusion transcripts}
\usage{
get_ppi_switches(hits_all, ppi)
}
\arguments{
\item{hits_all}{`data.frame` or `data.table` of event-level hits,
including `transcript_id_inc`, `transcript_id_exc`, and `pc_class`.}

\item{ppi}{PPI map (e.g. IntAct or STRING) containing columns
`ensembl_transcript_id` and `i.ensembl_transcript_id`.}
}
\value{
A `data.table` identical to `hits_all` with added columns:
\describe{
  \item{`inc_ppi`, `exc_ppi`}{Lists of partner transcripts unique to
    inclusion or exclusion isoforms.}
  \item{`n_inc_ppi`, `n_exc_ppi`}{Counts of gained/lost interactions.}
  \item{`n_ppi`}{Total PPI changes (sum of both directions).}
}
}
\description{
Identifies gained and lost protein protein interactions (PPIs)
between inclusion (INC) and exclusion (EXC) isoforms for each
splicing event.
}
\examples{
sample_frame <- data.frame(path = c(check_extdata_dir('rawData/control_S5/'),
                                    check_extdata_dir('rawData/control_S6/'),
                                    check_extdata_dir('rawData/control_S7/'),
                                    check_extdata_dir('rawData/control_S8/'),
                                    check_extdata_dir('rawData/case_S1/'),
                                    check_extdata_dir('rawData/case_S2/'),
                                    check_extdata_dir('rawData/case_S3/'),
                                    check_extdata_dir('rawData/case_S4/')),
                           sample_name  = c("S5", "S6", "S7", "S8", "S1", "S2", "S3", "S4"),
                           condition    = c("control", "control", "control", "control", "case",  "case",  "case",  "case"),
                           stringsAsFactors = FALSE)
hit_index <- get_hitindex(sample_frame)
res <- get_differential_inclusion(hit_index)
annotation_df <- get_annotation(load = "test")
matched <- get_matched_events_chunked(res, annotation_df$annotations, chunk_size = 2000)
x_seq <- attach_sequences(matched, annotation_df$sequences)
pairs <- get_pairs(x_seq, source="multi")
seq_compare <-compare_sequence_frame(pairs, annotation_df$annotations)

annotation_df <- get_annotation(load = 'test')
interpro_features <- get_protein_features(c("interpro"), annotations$annotations, timeout = 600, test = TRUE)
protein_feature_total <- get_comprehensive_annotations(list(interpro_features))

exon_features <- get_exon_features(annotation_df$annotations, protein_feature_total)

hits_domain <- get_domains(seq_compare, exon_features)

bg <- get_background(source = "hit_index",
                     input = sample_frame,
                     annotations = annotation_df$annotations,
                     protein_features = protein_feature_total)
ppidm <- get_ppidm(test=TRUE)

ppi <- get_isoform_interactions(protein_feature_total[ensembl_transcript_id \%in\%
                                                        annotation_df$annotations[, unique(transcript_id)]],
                                ppidm, save = FALSE, load_dir = '/projectnb2/evolution/zwakefield/proteinImpacts/', init = TRUE)


hits_final <- get_ppi_switches(hits_domain, ppi)
}
