% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_features.R
\name{get_protein_features}
\alias{get_protein_features}
\title{External function to fetch protein features from biomaRt}
\usage{
get_protein_features(
  biomaRt_databases = c("interpro", "mobidblite", "seg", "ncoils", "tmhmm", "signalp"),
  gtf_df,
  load_path = NULL,
  save_path = NULL,
  timeout = 600,
  test = FALSE
)
}
\arguments{
\item{biomaRt_databases}{choose what biomaRt attribute to access, defaulting
to interpro, mobidblite, seg, ncoils, tmhmm, signalp}

\item{load_path}{path to load prior protein features from}

\item{save_path}{path to save prior protein features from}

\item{timeout}{ability to extend timeout if biomaRt is not cooperating}

\item{test}{Logical; bool for whether to load from reduced test set}
}
\description{
External function to fetch protein features from biomaRt
}
\examples{
annotation_df <- get_annotation(load = "test")
interpro_features <- get_protein_features(c("interpro"), annotation_df$annotations, timeout = 600, test_set = TRUE)

}
