% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_features.R
\name{get_protein_features}
\alias{get_protein_features}
\title{External function to fetch protein features from biomaRt}
\usage{
get_protein_features(
  biomaRt_databases = c("interpro", "mobidblite", "seg", "ncoils", "tmhmm", "signalp"),
  gtf_df,
  load_path = NULL,
  save_path = NULL,
  timeout = 600,
  species = c("human", "mouse")[1],
  release = c(109, 115)[1],
  test = FALSE
)
}
\arguments{
\item{biomaRt_databases}{choose what biomaRt attribute to access, defaulting
to interpro, mobidblite, seg, ncoils, tmhmm, signalp}

\item{gtf_df}{annotations from get_annotation()}

\item{load_path}{path to load prior protein features from}

\item{save_path}{path to save prior protein features from}

\item{timeout}{ability to extend timeout if biomaRt is not cooperating}

\item{species}{Character string giving the Ensembl BioMart
dataset (default \code{"human"}). For mouse, use
\code{"mouse"}.}

\item{release}{release version from ensemble associated with the gencode
version provided in the get_annotation (This can be found in gencode, for
example in human: https://www.gencodegenes.org/human/releases.html)
example in mouse: https://www.gencodegenes.org/mouse/releases.html)}

\item{test}{Logical; bool for whether to load from reduced test set}
}
\value{
A `data.table` with one row per protein feature and transcript coupling
}
\description{
Here we also remove any duplicate and overlapping domains
We also add the genomic location to the name of the protein feature for
downstream safeguarding and precision. This is to prevent different occurences
of the same domain being called as the same in domain identification and
enrichment.
}
\examples{
annotation_df <- get_annotation(load = "test")
interpro_features <- get_protein_features(c("interpro"), annotation_df$annotations, timeout = 600, test = TRUE)
}
