% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_hit_index.R
\name{get_rmats_hit}
\alias{get_rmats_hit}
\title{Wrapper function to get both rmats and hit index cleanly}
\usage{
get_rmats_hit(
  sample_frame,
  event_types = c("MXE", "SE", "A3SS", "A5SS", "RI"),
  use = "JCEC",
  keep_annotated_first_last = TRUE
)
}
\arguments{
\item{sample_frame}{Data.frame with columns: \code{path}, \code{condition}, and \code{sample_name}.}

\item{event_types}{event types to load from rMATS}

\item{use}{Character scalar, one of \code{"JC"} or \code{"JCEC"}.}

\item{keep_annotated_first_last}{Logical; if TRUE, retain only annotated first/last exons and normalize PSI.}
}
\description{
Wrapper function to get both rmats and hit index cleanly
}
