% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_hit_index.R
\name{get_rmats_hit}
\alias{get_rmats_hit}
\title{Wrapper function to get both rmats and hit index cleanly}
\usage{
get_rmats_hit(
  sample_frame,
  event_types = c("ALE", "AFE", "MXE", "SE", "A3SS", "A5SS", "RI"),
  use = "JCEC",
  keep_annotated_first_last = TRUE
)
}
\arguments{
\item{sample_frame}{Data.frame with columns: \code{path}, \code{condition}, and \code{sample_name}.}

\item{event_types}{event types to load from rMATS}

\item{use}{Character scalar, one of \code{"JC"} or \code{"JCEC"}.}

\item{keep_annotated_first_last}{Logical; if TRUE, retain only annotated first/last exons and normalize PSI.}
}
\value{
a `data.table` for all the event types desired from the paths supplied
contains: event_id (unique id for event), event_type (AS event type), form
(INC/EXC), gene_id (ensembl id), strand, inc, exc (inclusion and exclusion coords)
inclusion reads, exclusion reads, psi, sample, condition, source file
}
\description{
Wrapper function to get both rmats and hit index cleanly
}
\examples{
sample_frame <- data.frame(path = c(check_extdata_dir('rawData/control_S5/'),
                                    check_extdata_dir('rawData/control_S6/'),
                                    check_extdata_dir('rawData/control_S7/'),
                                    check_extdata_dir('rawData/control_S8/'),
                                    check_extdata_dir('rawData/case_S1/'),
                                    check_extdata_dir('rawData/case_S2/'),
                                    check_extdata_dir('rawData/case_S3/'),
                                    check_extdata_dir('rawData/case_S4/')),
                           sample_name  = c("S5", "S6", "S7", "S8", "S1", "S2", "S3", "S4"),
                           condition    = c("control", "control", "control", "control", "case",  "case",  "case",  "case"),
                           stringsAsFactors = FALSE)
data <- get_rmats_hit(sample_frame, event_types = c("ALE", "AFE", "MXE", "SE", "A3SS", "A5SS", "RI"))
}
