% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_rmats.R
\name{get_rmats_post_di}
\alias{get_rmats_post_di}
\title{Import post-differential-inclusion rMATS results}
\usage{
get_rmats_post_di(input, event_type = NULL)
}
\arguments{
\item{input}{Either:
* a data.frame with columns `path`, `grp1`, `grp2`, `event_type`, or
* a data.frame of rMATS post-DI results.}

\item{event_type}{Optional event type when `input` contains a single
rMATS data.frame. Ignored when file metadata table is supplied.}
}
\value{
A `data.table` with columns:
\describe{
  \item{event_id}{unique event identifier}
  \item{event_type}{splicing event type}
  \item{form}{\code{"INC"} or \code{"EXC"}}
  \item{gene_id}{gene ID}
  \item{chr}{chromosome}
  \item{strand}{strand}
  \item{inc}{genomic coordinates of included segment(s)}
  \item{exc}{genomic coordinates of excluded segment(s)}
  \item{p.value}{rMATS p-value}
  \item{padj}{FDR}
  \item{delta_psi}{signed PSI change (+INC, -EXC)}
}
}
\description{
This function reads post-DI rMATS results and converts them into
the standardized SpliceImpactR long format with one row per
event x (INC/EXC) form.

Input can be:
* a data.frame with columns `path`, `grp1`, `grp2`, `event_type`,
  in which case each file is read and processed; or
* a single rMATS results data.frame, in which case `event_type`
  must be supplied.

For each event, the function constructs paired INC and EXC entries:
* `inc` contains genomic segments included in the form
* `exc` contains the excluded segment(s)
* `delta_psi`, `p.value`, and `padj` are assigned using
  the rMATS-reported values

Event IDs are automatically generated (event_type:N) if not supplied
}
\examples{
# # Multiple files
# input <- data.frame(
#   path = c('/path/A3SS.MATS.JC.txt', '/path2/A5SS.MATS.JC.txt'),
#   grp1 = c("WT","WT"),
#   grp2 = c("KO","KO"),
#   event_type = c("A3SS", "A5SS")
# )
# res <- get_rmats_post_di(meta)

# Single rMATS table already loaded as df
df <- data.frame(
  ID = 1L,
  GeneID = "ENSG00000182871",
  geneSymbol = "COL18A1",
  chr = "chr21",
  strand = "+",
  longExonStart_0base = 45505834L,
  longExonEnd = 45505966L,
  shortES = 45505837L,
  shortEE = 45505966L,
  flankingES = 45505357L,
  flankingEE = 45505431L,
  ID.2 = 2L,
  IJC_SAMPLE_1 = "4,1,0",
  SJC_SAMPLE_1 = "9,12,3",
  IJC_SAMPLE_2 = "0,4,5",
  SJC_SAMPLE_2 = "11,15,15",
  IncFormLen = 52L,
  SkipFormLen = 49L,
  PValue = 0.6967562,
  FDR = 1,
  IncLevel1 = "0.295,0.073,0.0",
  IncLevel2 = "0.0,0.201,0.239",
  IncLevelDifference = -0.024,
  stringsAsFactors = FALSE
)
res2 <- get_rmats_post_di(df, event_type = "A3SS")
print(res2)
}
