% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{getAnnotation}
\alias{getAnnotation}
\title{wrapper for setup_gtf, extract annotation information}
\usage{
getAnnotation(biomart_data)
}
\arguments{
\item{biomart_data}{output from setup_biomart}
}
\value{
annotated gtf and others from setup_gtf
}
\description{
wrapper for setup_gtf, extract annotation information
}
\examples{

pdir <- system.file("extdata", package = "SpliceImpactR")
dataDirectory <- paste0(pdir, "/")

setup_gtf_exon_data <- readr::read_csv(paste0(dataDirectory, "biomart_setup_gtf_exon_data.csv"))
ptg_init <- readr::read_csv(paste0(dataDirectory, "biomart_ptg_init.csv"))
transcript_data <- readr::read_csv(paste0(dataDirectory, "biomart_transcript_data.csv"))

biomart_data_sample <- list(setup_gtf_exon_data = setup_gtf_exon_data,
                            ptg_init = ptg_init,
                            transcript_data = transcript_data)
gtf <- getAnnotation(biomart_data = biomart_data_sample)

}
