% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBackground.R
\name{getBackground}
\alias{getBackground}
\title{getBackground set of transcripts in samples}
\usage{
getBackground(
  input,
  mOverlap,
  cores,
  exon_type,
  output_location = NULL,
  gtf,
  translations
)
}
\arguments{
\item{input}{the paths containing the .exon files (/a/b/c if the file is /a/b/c.exon)}

\item{mOverlap}{overlap to identify a match to annotation}

\item{cores}{number of cores to use}

\item{exon_type}{aRNAp event type}

\item{output_location}{location to make background directory}

\item{gtf}{full output from getAnnotation}

\item{translations}{from getTranslations}
}
\value{
matched : matched transcripts dataframe, bed : bed file of the matched transcripts
proBed : output for further functions with protein code and protein info,
proFast : fasta file of proteins identified in proBed
}
\description{
getBackground set of transcripts in samples
}
\examples{

pdir <- system.file("extdata", package="SpliceImpactR")
dataDirectory <- paste0(pdir, "/")
test_group <- paste0(dataDirectory, "rawData/", c("test1","test2", "test3"))
control_group <- paste0(dataDirectory, "rawData/", c("control1", "control2", "control3"))
gtf_sample <- list(gtf = readr::read_csv(paste0(dataDirectory, "gtf_limited.csv")),
            transcript_gtf = readr::read_csv(paste0(dataDirectory, "transcript_gtf_limited.csv")))
translations_sample <- readr::read_lines(paste0(dataDirectory, "translations_limited.csv"))

bg <- getBackground(input=c(test_group, control_group),
                    mOverlap = 0.1,
                    cores = 1,
                    exon_type = "AFE",
                    output_location = NULL, gtf_sample, translations_sample)

}
