% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getForeground.R
\name{getForeground}
\alias{getForeground}
\title{get Foreground differentially included transcripts in samples}
\usage{
getForeground(
  input,
  test_names,
  control_names,
  thresh,
  fdr,
  mOverlap,
  exon_type,
  pdir,
  output_location,
  cores = 1,
  gtf
)
}
\arguments{
\item{input}{diExon info from the differentially included functions}

\item{thresh}{delta psi threshold to filter}

\item{fdr}{adj p to filter for}

\item{mOverlap}{overlap to identify a match to annotation}

\item{exon_type}{placeholder for other functions}

\item{pdir}{location of the package}

\item{output_location}{location to make background directory}

\item{gtf}{gtf dataframe from setup_gtf}
}
\value{
matched : matched transcripts dataframe, bed : bed file of the matched transcripts
proBed : output for further functions with protein code and protein info,
proFast : fasta file of proteins identified in proBed
}
\description{
get Foreground differentially included transcripts in samples
}
