% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfxnlASoutcome.R
\name{getfxnlASoutcome}
\alias{getfxnlASoutcome}
\title{Full run of the pipeline for a single alternative splicing event type}
\usage{
getfxnlASoutcome(
  output_location,
  test_group,
  control_group,
  data_df,
  exon_type,
  cutoff = 0.1,
  outlier_handle = "4/n",
  cores = 1,
  tti_location = "",
  full_pipe = TRUE,
  bg = NA,
  mOverlap = 0.05,
  gtf
)
}
\arguments{
\item{output_location}{location to make  directory}

\item{test_group}{the paths of one phenotype}

\item{control_group}{the paths of one phenotype}

\item{data_df}{dataframe with sample paths, type, and phenotype}

\item{exon_type}{type of exon being investigated}

\item{cutoff}{cutoff for significance of differential inclusion}

\item{tti_location}{location of previuously made tti or ""}

\item{full_pipe}{if TRUE, doesn't save output to R, only writes to output_location}

\item{bg}{whether bg preran or needs making}

\item{mOverlap}{overlap to identify a match to annotation}

\item{gtf}{gtf data frame from setup_gtf}
}
\value{
nothing or all output from pipeline
}
\description{
Full run of the pipeline for a single alternative splicing event type
}
