% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHitCompare.R
\name{getHitCompare}
\alias{getHitCompare}
\title{get paired as events / exons}
\usage{
getHitCompare(data_df, output_location, threshold = 0.25)
}
\arguments{
\item{data_df}{input with control and test noted}

\item{output_location}{output location for plots}
}
\value{
figure showing proximal / distal shift and table of results
}
\description{
get paired as events / exons
}
