% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHitCompare.R
\name{getHitCompare}
\alias{getHitCompare}
\title{Perform analysis on how the HIT Index value is changing on an exon-basis}
\usage{
getHitCompare(data_df, output_location, threshold = 0.1)
}
\arguments{
\item{data_df}{input with control and test noted}

\item{output_location}{output location for plots}

\item{threshold}{for NA in HIT Index}
}
\value{
figures about HIT Index and csv for differentially used exons (diff exon types)
}
\description{
Perform analysis on how the HIT Index value is changing on an exon-basis
}
\examples{

pdir <- system.file("extdata", package="SpliceImpactR")
dataDirectory <- paste0(pdir, "/rawData/")
test_group <- paste0(dataDirectory, c("test1", "test2", "test3"))
control_group <- paste0(dataDirectory, c("control1", "control2", "control3"))
data_df <- data.frame(
    sample_names = c(control_group, test_group),
    phenotype_names = c(
      rep("control", length(control_group)),
      rep("test", length(test_group))
     ),
   stringsAsFactors = FALSE
  )
  data_df$utc <- "control"
  data_df$utc[data_df$phenotype_names == unique(data_df$phenotype_names)[2]] <- "test"
compareHIT <- getHitCompare(data_df, output_location = NULL, threshold = .1)

}
