% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLengthComparison.R
\name{getLengthComparison}
\alias{getLengthComparison}
\title{Get comparison between length across isoform swaps due to AS}
\usage{
getLengthComparison(data_df, paired_df, output_location)
}
\arguments{
\item{data_df}{df with paths, phenotype names, class}

\item{paired_df}{from getPaired df}

\item{output_location}{location where everything is being saved}
}
\value{
1 combined plot.
}
\description{
Uses the paired results to identify any potential changes in lengths across phenotype
}
\examples{
getLengthComparison(data_df, paired_df from pfg, output_path)

}
