% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameShiftDetector.R
\name{getNextOverlap}
\alias{getNextOverlap}
\title{get next downstream exon for MXE, SE}
\usage{
getNextOverlap(df1, df2, e1, e2, eld)
}
\arguments{
\item{df1}{info related to first transcript}

\item{df2}{info related to second transcript}

\item{e1}{first exon}

\item{e2}{second exon}

\item{eld}{exon_length_df from previous function}
}
\value{
either new exon set for limiting at least one df
}
\description{
get next downstream exon for MXE, SE
}
\keyword{internal}
