% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverviewComparison.R
\name{getOverviewComparison}
\alias{getOverviewComparison}
\title{Get comparison between phenotypes: count of AS event, count of AS event per gene, ECDF compare and save to Foreground subdir}
\usage{
getOverviewComparison(
  data_df,
  exon_type,
  output_location,
  plot = T,
  minReads = 10
)
}
\arguments{
\item{exon_type}{type of AS being investigated}

\item{output_location}{location where everything is being saved}

\item{control_names}{paths of control samples}

\item{test_names}{paths of test samples}
}
\value{
4 individual plots and 1 combined plot.
}
\description{
Get comparison between phenotypes: count of AS event, count of AS event per gene, ECDF compare and save to Foreground subdir
}
\examples{
getOverviewComparison(c("path_control1", "path_control2"), c("path_test1", "path_test2"), "AFE", "path_to_output")
}
