% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverviewComparison.R
\name{getOverviewComparison}
\alias{getOverviewComparison}
\title{Get comparison between phenotypes: count of AS event, count of AS event per gene, ECDF compare and save to Foreground subdir}
\usage{
getOverviewComparison(data_df, exon_type, output_location, minReads = 10)
}
\arguments{
\item{data_df}{data_df made in specific format, used elsewhere}

\item{exon_type}{type of AS being investigated}

\item{output_location}{location where everything is being saved, NULL if not desired}

\item{minReads}{to use while filtering}
}
\value{
4 individual plots and 1 combined plot.
}
\description{
Get comparison between phenotypes: count of AS event, count of AS event per gene, ECDF compare and save to Foreground subdir
}
\examples{

 pdir <- system.file("extdata", package="SpliceImpactR")
 dataDirectory <- paste0(pdir, "/rawData/")
 test_group <- paste0(dataDirectory, c("test1","test2", "test3"))
 control_group <- paste0(dataDirectory, c("control1", "control2", "control3"))
 data_df <- data.frame(sample_names = c(control_group, test_group),
                       phenotype_names = c(rep("control", length(control_group)),
                       rep("test", length(test_group))))

 data_df$utc <- "control"
 data_df$utc[data_df$phenotype_names == unique(data_df$phenotype_names)[2]] <- "test"
 overview <- getOverviewComparison(data_df, "AFE", output_location = NULL, minReads = 10)

}
