% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPaired.R
\name{getPaired}
\alias{getPaired}
\title{get paired AS events / exons and identify similarities/differences through sequence alignment + identify frame shifts}
\usage{
getPaired(
  foreground,
  et,
  nucleotides,
  newGTF,
  cores = 1,
  output_location = NULL,
  saveAlignments = TRUE,
  exon_data
)
}
\arguments{
\item{foreground}{proBed from getForeground}

\item{et}{exon_type being investigated}

\item{nucleotides}{made from setup get_c_nucs}

\item{newGTF}{from setup get_gtf}

\item{cores}{number of requested cores}

\item{output_location}{location to make background directory}

\item{saveAlignments}{bool to save alignment pdfs}

\item{exon_data}{from biomart_data}
}
\value{
figures and dataframes with paired data
}
\description{
get paired AS events / exons and identify similarities/differences through sequence alignment + identify frame shifts
}
\examples{

pdir <- system.file("extdata", package="SpliceImpactR")
dataDirectory <- paste0(pdir, "/")
test_group <- paste0(dataDirectory, "rawData/", c("test1","test2", "test3"))
control_group <- paste0(dataDirectory, "rawData/", c("control1", "control2", "control3"))
transDF <- readr::read_csv(paste0(dataDirectory, "transcripts_limited_transDF.csv"))
c_trans <- readr::read_lines(paste0(dataDirectory, "transcripts_limited_c_trans.csv"))

transcripts_sample <- list(transDF = transDF,
                           c_trans = c_trans)

gtf_sample <- list(gtf = readr::read_csv(paste0(dataDirectory, "gtf_limited.csv")),
            transcript_gtf = readr::read_csv(paste0(dataDirectory, "transcript_gtf_limited.csv")))
translations_sample <- readr::read_lines(paste0(dataDirectory, "translations_limited.csv"))
biomart_data_sample <- readr::read_csv(paste0(dataDirectory, "biomart_data_sample.csv"))


result <- differential_inclusion_HITindex(test_names = test_group,
                                          control_names = control_group,
                                          et = "AFE",
                                          outlier_threshold = "Inf",
                                          minReads = 10,
                                          min_prop_samples = 0,
                                          chosen_method = "qbGLM"
                                          )

fg <- getForeground(input = result,
                            test_names = test_group,
                            control_names = control_group,
                            thresh = .1,
                            fdr = .05,
                            mOverlap = .1,
                            exon_type = "AFE",
                            output_location = NULL,
                            cores = 1,
                            gtf = gtf_sample,
                            max_zero_prop = 1,
                            min_prop_samples = 0,
                            translations = translations_sample)
library(msa)
pfg <- getPaired(foreground = fg$proBed,
          et = "AFE",
          nucleotides = transcripts_sample,
          newGTF = gtf_sample,
          cores = 1,
          output_location = NULL,
          saveAlignments = FALSE,
          exon_data = biomart_data_sample)

}
