% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{HITmatcher}
\alias{HITmatcher}
\title{specific matcher for HIT}
\usage{
HITmatcher(i, redExon, gtf_filtered, minOverlap, protein_coding_transcripts)
}
\value{
matched transcript rownumber
}
\description{
specific matcher for HIT
}
