% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{HITmatcher}
\alias{HITmatcher}
\title{specific matcher for HIT}
\usage{
HITmatcher(i, redExon, gtf_filtered, minOverlap, protein_coding_transcripts)
}
\arguments{
\item{i}{index of redExon}

\item{redExon}{dataframe generated initially in getForeground}

\item{gtf_filtered}{gtf filtered for gene}

\item{minOverlap}{minimum overlap to count as same exon}

\item{protein_coding_transcripts}{protein coding transcripts}
}
\value{
matched transcript rownumber
}
\description{
specific matcher for HIT
}
\keyword{internal}
