% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTTI.R
\name{init_ddi}
\alias{init_ddi}
\title{initiates tti network for entire genome, can be very time consumng}
\usage{
init_ddi(
  pdir,
  output_location = NULL,
  ppidm_class = c("Gold_Standard", "Gold", "Silver", "Bronze")[1],
  removeDups = TRUE,
  cores = 1,
  pfam_data
)
}
\arguments{
\item{pdir}{the directory of the package}

\item{output_location}{location to make output}

\item{ppidm_class}{threshold of ppidm sets to use}

\item{removeDups}{takes a longer time to init, but faster down the line and saves a lot of memory to remove duplcate vertices}

\item{cores}{number of requested cores}

\item{pfam_data}{from biomart_data, pfam_data}
}
\value{
overall tti network in both graph and edgelist form
}
\description{
initiates tti network for entire genome, can be very time consumng
}
\examples{

pdir <- system.file("extdata", package="SpliceImpactR")
dataDirectory <- paste0(pdir, "/")

ip <- readr::read_csv(paste0(dataDirectory, "biomart_ip.csv"))
code_regions <- readr::read_csv(paste0(dataDirectory, "biomart_code_regions.csv"))
pfam_exon_level <- readr::read_csv(paste0(dataDirectory, "biomart_pfam_exon_level.csv"))
fsd_exon_data <- readr::read_csv(paste0(dataDirectory, "biomart_data_sample.csv"))
pfam_data = readr::read_csv(paste0(dataDirectory, "biomart_pfam_exon.csv"))
biomart_data_sample <- list(ip = ip,
                     code_regions = code_regions,
                     fsd_exon_data = fsd_exon_data,
                     pfam_exon_level = pfam_exon_level,
                     pfam_data = pfam_data)

initDDI <- init_ddi(pdir = dataDirectory,
                    output_location = NULL,
                    ppidm_class = c("Gold_Standard", "Gold", "Silver", "Bronze")[1],
                    removeDups = TRUE,
                    cores = 1,
                    pfam_data = biomart_data_sample$pfam_data)

}
