% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppi.R
\name{ipr_to_pfam}
\alias{ipr_to_pfam}
\title{Convert InterPro IDs to PFAM IDs}
\usage{
ipr_to_pfam(ipr_ids)
}
\arguments{
\item{ipr_ids}{Character vector of InterPro IDs (for example, `"IPR000719"`).}
}
\value{
Character vector of PFAM IDs mapped from `ipr_ids`.
}
\description{
Convert InterPro IDs to PFAM IDs
}
\keyword{internal}
