% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{matcher}
\alias{matcher}
\title{matches the given locations to annotations and transcripts}
\usage{
matcher(
  ex_type,
  background = FALSE,
  cores = 1,
  redExon,
  minOverlap = 0.05,
  gtf,
  gtf_transcripts
)
}
\arguments{
\item{ex_type}{type of exon being queried}

\item{background}{whether making background or not}

\item{cores}{number of requested cores}

\item{redExon}{input of simplified exon coords internally}

\item{minOverlap}{minimum overlap to classify as matched to annotation}

\item{gtf}{dataframe from setup_gtf}

\item{gtf_transcript}{dataframe from setup_gtf}
}
\value{
figures and dataframes with paired data
}
\description{
matches the given locations to annotations and transcripts
}
\keyword{internal}
