% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_differential_inclusion.R
\name{plot_di_volcano_dt}
\alias{plot_di_volcano_dt}
\title{Volcano plot for differential inclusion results}
\usage{
plot_di_volcano_dt(di, padj_thr = 0.05, dpsi_thr = 0.1)
}
\arguments{
\item{di}{A `data.frame` or `data.table` containing at least
`delta_psi` and `padj` columns (optionally `event_type`).}

\item{padj_thr}{Numeric. Adjusted p-value threshold (default `0.05`).}

\item{dpsi_thr}{Numeric. Absolute deltaPSI threshold (default `0.1`).}
}
\value{
A `ggplot2` object showing differential inclusion significance.
}
\description{
Generates a volcano plot of deltaPSI vs. -log10(FDR) highlighting significant events.
}
\details{
Significant sites are colored in `deeppink4`; nonsignificant sites are
shown in light grey. Dashed and dotted lines indicate deltaPSI and FDR thresholds.
}
\examples{
sample_frame <- data.frame(path = c(check_extdata_dir('rawData/control_S5/'),
                                    check_extdata_dir('rawData/control_S6/'),
                                    check_extdata_dir('rawData/control_S7/'),
                                    check_extdata_dir('rawData/control_S8/'),
                                    check_extdata_dir('rawData/case_S1/'),
                                    check_extdata_dir('rawData/case_S2/'),
                                    check_extdata_dir('rawData/case_S3/'),
                                    check_extdata_dir('rawData/case_S4/')),
                           sample_name  = c("S5", "S6", "S7", "S8", "S1", "S2", "S3", "S4"),
                           condition    = c("control", "control", "control", "control", "case",  "case",  "case",  "case"),
                           stringsAsFactors = FALSE)
hit_index <- get_hitindex(sample_frame)
res <- get_differential_inclusion(hit_index)
plot_di_volcano_dt(res)

}
