% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diHelpers.R
\name{qualityFilter}
\alias{qualityFilter}
\title{filter using minimums for identifying events in control/test and max proportion of zeros allowed}
\usage{
qualityFilter(df, nT, nC, min_prop_samples = 0.5, max_zero_prop = 0.5)
}
\arguments{
\item{nT}{number of test samples}

\item{nC}{number of control samples}

\item{min_prop_samples}{min proportion of samples needed to not be counted as outliers}

\item{max_zero_prop}{max number of zeros in to count an exon}

\item{final_data}{from diHIT or diRMATS passed through diColor}
}
\value{
a filtered di inclusion dataframe
}
\description{
filter using minimums for identifying events in control/test and max proportion of zeros allowed
}
\keyword{internal}
