% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{RImatcher}
\alias{RImatcher}
\title{specific matcher for RI}
\usage{
RImatcher(
  i,
  below_thresh = 0.2,
  redExon,
  minOverlap = 0.05,
  gtf_transcripts,
  gtf_exons,
  protein_coding_transcripts,
  transcript_starts
)
}
\value{
matched transcript rownumber
}
\description{
specific matcher for RI
}
\keyword{internal}
