% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gtf.R
\name{setup_gtf}
\alias{setup_gtf}
\title{setting up gtf with exon labels, first, internal, last, hybrid, etc}
\usage{
setup_gtf()
}
\value{
annotated gtf and various hybrid info. also gene to transcript to protein naming dataframe
}
\description{
setting up gtf with exon labels, first, internal, last, hybrid, etc
}
\keyword{internal}
