% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gtf.R
\name{setupBiomart}
\alias{setupBiomart}
\title{import / load + save biomaRt data}
\usage{
setupBiomart(save_location)
}
\arguments{
\item{save_location}{path to save output}
}
\value{
various biomaRt results for use downstream in pipeline
}
\description{
import / load + save biomaRt data
}
\examples{
\donttest{
library(biomaRt)
new_config <- httr::config(ssl_verifypeer = FALSE)
httr::set_config(new_config, override = FALSE)
ensembl <- biomaRt::useEnsembl(biomart = "ensembl",
                               dataset = "hsapiens_gene_ensembl",
                               mirror = 'useast')
pdir <- system.file("extdata", package = "SpliceImpactR")
dataDirectory <- paste0(pdir, "/")
biomart_data <- setupBiomart(dataDirectory)
}
}
